+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Use Promise.allSettled() to ensure we wait for all loads to complete before marking state as not loading any more
 * Move `minCharsForLanguageDetect` and `maxContentPreviewChars` to `TootsConfig`
 
 ### v0.69.11